requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
camelot-py
opencv-python-headless
pytesseract
Pillow
pandas
openpyxl
tabulate
openai
python-dotenv
